Creates skeletons for Results, Upcoming and MyFixtures pages
