+++ front-end/Components/MyFixturesPage.js
@@ -1,19 +1,49 @@
-import { StyleSheet } from "react-native";
-import { Card, Text } from "react-native-paper";
+import { StyleSheet, ScrollView } from "react-native";
+import FixtureCard from "./FixtureCard";
+import MatchdayContainer from "./MatchdayContainer";
+import MyFixtureUI from "./MyFixtureUI";
 
 export default function MyFixturesPage() {
+    const futureFixture = {
+        fixture_id: 1,
+        match_status: "planned",
+        team1_id: 1,
+        team1_name: "Leicester Wolves",
+        team2_id: 2,
+        team2_name: "Old Bags",
+        venue_id: 3,
+        venue_name: "Alfie Fenables Centre",
+        match_date: "11/5/24",
+        start_time: 13.0,
+        division: "Division 1",
+    };
+
     const styles = StyleSheet.create({
-        container: {
+        scroll: {
             flex: 1,
             alignItems: "center",
-            justifyContent: "center",
-            margin: 20,
+            gap: 5,
+            padding: 5,
         },
     });
 
     return (
-        <Card style={styles.container}>
-            <Text>My Fixtures Page</Text>
-        </Card>
+        <ScrollView contentStyle={styles.scroll}>
+            <MatchdayContainer date={futureFixture.match_date}>
+                <FixtureCard fixture={futureFixture}>
+                    <MyFixtureUI />
+                </FixtureCard>
+            </MatchdayContainer>
+            <MatchdayContainer date={futureFixture.match_date}>
+                <FixtureCard fixture={futureFixture}>
+                    <MyFixtureUI />
+                </FixtureCard>
+            </MatchdayContainer>
+            <MatchdayContainer date={futureFixture.match_date}>
+                <FixtureCard fixture={futureFixture}>
+                    <MyFixtureUI />
+                </FixtureCard>
+            </MatchdayContainer>
+        </ScrollView>
     );
 }
